Tramp: only one file-notify process per connection/flags
The previous handlers in tramp-sh.el for the file-notify functions
were launching a single remote process per watched file, which quickly
results in too many open connections.

To alleviate this problem, this patch modifies the Tramp handler to
maintain a list of files which are requested to be watched for the
connection. It then launches a single remote process on the connection
for each combination of flags. When the list is updated, the watcher
process is relaunched with the new list of files.

The debounce is added so that recurrent calls to file-notify-add-watch
don't cause a slowdown. For example, eglot calls this function with
every directory in the project in order to handle the
workspace/didChangeWatchedFiles request from LSP servers.